CXSPA-1301: Move unit-order display tests to common test file
